+++ app/src/main/java/com/lastaoutdoor/lasta/ui/screen/discover/DiscoverScreen.kt
@@ -61,9 +61,9 @@ import com.lastaoutdoor.lasta.ui.components.WeatherReportSmall
 import com.lastaoutdoor.lasta.ui.screen.discover.components.ModalUpperSheet
 import com.lastaoutdoor.lasta.ui.screen.discover.components.RangeSearchComposable
 import com.lastaoutdoor.lasta.ui.screen.map.MapScreen
+import com.lastaoutdoor.lasta.utils.OrderingBy
 import com.lastaoutdoor.lasta.viewmodel.DiscoverDisplayType
 import com.lastaoutdoor.lasta.viewmodel.MapState
-import com.lastaoutdoor.lasta.viewmodel.OrderingBy
 
 @Composable
 fun DiscoverScreen(
